move default file logic under track files
