+++ test/generator/files/track_files_test.rb
@@ -60,30 +60,23 @@ module Generator
         assert_instance_of MinitestTestsFile, subject.minitest_tests
       end
 
-      def test_track_tests_template
+      def test_tests_template
         subject = TestTrackFiles.new
-        assert_instance_of TestsTemplateFile, subject.track_tests_template
+        assert_instance_of TestsTemplateFile, subject.tests_template
       end
-    end
 
-    class DefaultFilesTest < Minitest::Test
-      FixturePaths = Paths.new(
-        metadata: 'test/fixtures/metadata',
-        track: 'test/fixtures/xruby'
-      )
-
-      class TestDefaultFiles
+      class TestTrackFilesUseDefault
         def initialize
           @paths = FixturePaths
-          @exercise_name = 'alpha'
+          @exercise_name = 'notemplate'
         end
         attr_reader :paths, :exercise_name
-        include DefaultFiles
+        include TrackFiles
       end
 
       def test_default_tests_template
-        subject = TestDefaultFiles.new
-        assert_instance_of TestsTemplateFile, subject.default_tests_template
+        subject = TestTrackFiles.new
+        assert_instance_of TestsTemplateFile, subject.tests_template
       end
     end
